// pslpwcl/src/app/modules/pslpapi/model/indirizzoStradario.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IndirizzoStradario { 
    cap?: string;
    circoscrizione?: string;
    civicoNumero?: string;
    civicoSub?: string;
    codIstat?: string;
    comune?: string;
    descrizione?: string;
    localita?: string;
    nomeVia?: string;
    provincia?: string;
    siglaProvincia?: string;
    tipoVia?: string;
}